Make fetchTransactions throwing and move provider out of AppDependencies default
  - TransactionsProviderProtocol.fetchTransactions now throws
  - ActivityViewModel handles the error case via LoadingState.error
  - AppDependencies no longer hardcodes InMemoryTransactionsProvider as a default; caller (BudgetTrackerApp) passes it explicitly
  - Transaction.amount changed from Float to Decimal
  - Add CLAUDE.md

+++ BudgetTracker/Common/Models/Transaction.swift
@@ -2,7 +2,7 @@ import Foundation
 
 struct Transaction: Hashable, Identifiable {
     let id: String
-    let amount: Float
+    let amount: Decimal
     let vendor: String
     let category: Category
     let date: Date
